Stop using a ChannelDownstreamHandler in the RegionClient.
When sending an RPC out to the wire, the code used to call:
  Channels.write(chan, rpc);
Then Netty would create a DownstreamMessageEvent wrapping the HBaseRpc,
and hand it to the RegionClient corresponding to this channel.
Then HBaseRpc would then get serialized into a ChannelBuffer, which
would then be passed on downstream (after being wrapped into another
new DownstreamMessageEvent) to the ChannelSink.

The new code instead serializes the HBaseRpc into a ChannelBuffer
before giving it to Netty, which will pass it straight to the
ChannelSink (after wrapping it in a DownstreamMessageEvent).  This
results in a smaller code path before we actually hit the wire and
fewer objects allocated on each write (so less work for the GC).

Change-Id: I0828f6443bf8c4d9e570e0c78720187b1543bdd2