Switch from LGPLv3+ to BSD license.
For the records, I wanted to use the LGPL, I think it's a fair license.
Unfortunately, the ASF doesn't allow inclusion of LGPL code in their
projects, and many seemed concerned about the "virality" of the LGPL.
I don't approve this change, and I tried to resist it, but due to
popular demand, this software package is switching to the BSD license.

Change-Id: Ice53f20c9927a708dc1805a2e1af3fa17be14125