Fix the retry logic for the multiPut RPC.
When a multiPut partially fails, the code builds another multiPut to
retry all the edits that are known to have potentially failed.  But
the new multiPut RPC that was built always had its `size' attribute
set to 0.

Also, MultiPutRequest had an `attempt' attribute that was
unintentionally shadowing the `attempt' attribute in the parent class
(HBaseRpc).  This was confusing and error-prone, so the attribute got
renamed as `retrycnt'.

MultiPutRequest now overrides toString in order to provide a meaningful
return value (which is particularly important when logging failed puts).

Change-Id: Ie0cb550c8c3b02d71cb2b3028c8e52776a28fc6b