Fix the comparator of .META. row keys to properly handle keys with commas.
A .META. row key looks like this:
    table,start_key,timestamp[.md5hash.]
But the code in RegionNameCmp failed to take into account that
`start_key' could contain an arbitrary number of commas.  The fix
is straightforward: start looking for the 2nd comma from the end.

This consequence of this bug was that sometimes, HBaseClient was unable
to execute RPCs for row keys containing a comma in them as it would find
the wrong entry in its region cache.

Change-Id: Ida8e608848bb65906183bfacae1da681efc942a4