Reduce lock contention on ip2client's lock.
When a new connection was opened to a RegionServer, the code
was unnecessarily holding the lock on the ip2client map while
configuring and connecting the channel.  I actually observed
contention on this lock because of that, even though connect()
is non-blocking and returns quickly.  The new code is slightly
awkward / less natural to read, but it's worth it.

Change-Id: I4482c49c60f50c9a5d6911a2154119755c4d5a77