Behave better when the znode of -ROOT- doesn't exist.
The code was closing the connection to ZK and sleeping 1s before
re-connecting to ZK and trying again.  Now the connection is kept
open and the code simply retries to read the znode again after 1s,
if the ZK session is still valid.

The ideal would be to setup a watch in ZK to wait until the znode
is created or updated.  This will be implemented later.

Change-Id: I305a24efee21f6226616fa28b9594e7fa32b925a