Fix the region name comparator and cache invalidation logic.
The region name comparator wasn't behaving properly when a table had a
name that was a prefix of another one (e.g. "mytable" and "mytablefoo").
When comparing two region names, such as:
  mytable,somekey,1234567890
  mytablefoo,otherkey,1234567890
It was returning the difference between first two different bytes (so
in this case, ',' - 'f' = 44 - 102 = -58).  This bug was non impacting
because ',' (=44) is less than every ASCII character typically used in
table names (typically [-_a-zA-Z0-9], which are all >44).

The cache invalidation logic wasn't properly clearing the regions_cache
when invalidating the last region of a table (the one with an empty stop
key).  It was actually working for all tables except the first one to
appear in the regions_cache (typically .META., unless all .META. regions
have already been invalidated from the cache).  The new logic uses a
different approach to locate the last region of a table in the cache
that also works for the first table.

Change-Id: I78e9ad0eb830f8a3316bf6fececb54ee63f6e246