Add missing synchronization.
client2regions is a map that associates each client connected to a
RegionServer to a list of regions that are known to be served by this
RegionServer.  The map itself is thread-safe, but the lists that were
added to it are not thread-safe, and they were accessed without any
synchronization.  Although there was a potential for race conditions,
those were extremely rare since this map is rarely accessed (only after
successful .META. lookups and after disconnecting from a RegionServer).

Change-Id: I392e5e2a60c9f972eee8891651ea4125f57f28e2