Fix the disconnection handling code.
The code was racy and could lead to lost RPCs or "deadlocks" when
getting unexpectedly disconnected from a RegionServer.  This was
particularly troublesome when HBase restarted (all of it).  When
HBase starts, it first brings online the -ROOT- region.  When the
client reads it to find where .META. is, it may find a stale entry
if the master hasn't updated it already.  Surprisingly, this happened
very frequently during restarts as HBase seems slow to start up.
Similarly, once -ROOT- was updated to point to .META. correctly,
the client could read stale .META. entries while the master was in
the process of bringing back up the regions and updating .META.

Under those circumstances, the client was frequently reading stale
ip:port information and getting a ConnectException.  The code that
was invalidating known information about the client wasn't doing the
right thing in this case, it wasn't invalidating anything, as it
made the wrong assumption that a client that had never been successfully
connected couldn't possibly be in the caches.

The new code prioritizes the invalidation of the reference to the
-ROOT- region when we get disconnected from it.  It also prioritizes
invalidation of our local cache of .META..

Because HBase is slow to start, this change is also increasing the
maximum number of attempts to send an RPC from 4 to 10.  In practice,
it seems that this is still insufficient, although hopefully with the
new master HBase will be able to start faster.  This means that during
an HBase restart, RPCs are likely to fail with a "Too many attempts"
exception, which is acceptable since HBase is indeed unavailable during
this period of time, and the application must be able to ride over such
transient HBase outages.  I'm wary of allow more retry attempts,
although this should be configurable (right now it's not).

Change-Id: I1cffd7bd22d3e6206af8997f5a29983de15dd296