Improve handling of NotServingRegionException slightly.
The code handling NotServingRegionException was suboptimal.  If the
RPC was on its last attempt, the code was making the RPC fail without
invaliding the local META cache.  Also, error reporting wasn't as
good as it could have been.  When an RPC fails because it was already
retried too many times, it would be good to know what the cause of
the last failure was (even though NotServingRegionException is the
most likely cause, swallowing the exception leads to sub-par error
reporting).

The code handling NSRE is still way suboptimal and doesn't always do
the right thing.  It'll be rewritten to be correct soon.

Change-Id: I4180ad007066684df5f0d9a8a16e883b5e83ed3a