Improve error reporting when remote exceptions occur.
Embed the HBaseRpc instance that caused the remote exception inside the
exception itself, and embed it in the message carried by the exception.
No longer pretends that NotServingRegionException can tell you which
region caused the exception, because we don't have this information
unless we manually try to parse the string of the exception returned
by the server, which we don't do because it's brittle at best.

Incompatible API changes: the following public method no longer exists:
  NotServingRegionException.getRegion()
Since async HBase is still beta software, I'm taking the liberty to
remove this from API without starting to carry the @Deprecated burden.

Change-Id: I8794d648723db609a6dbcf47a031db20270c4632