Add assertions as a safety net.
Refuse to encode an outgoing RPC if it doesn't have a Deferred already.
This happens when the caller didn't call getDeferred() on the RPC before
sending it out to the wire, which indicates a programming bug in the
HBase async library.  I tripped on this bug a few times already so this
assertion will help immediately catch it during development.

Refuse to allow an RPC to be given a RegionInfo if this RPC isn't going
to a specific table / key.  This should never happen and it will be even
more of a problem with the new NSRE handling code, so the assertion will
help catch this class of bugs early during development.
This requires changing the ReleaseRequest for RowLocks slightly because
it was specifying a region without a table / key pair, as those are not
needed / used for this RPC.

Change-Id: Id3264f9d1c735f59e10d9c0c3793375814dfdc0c