Completely re-implement the write path (PutRequest and such).
Warning: This change contains incompatible API changes (at the end).

The existing write path was too complicated and didn't fit well in the
rest of the code.  It had its own very special (and very long) code path
to send RPCs to the wire, which made handling those RPCs internally a
PITA.  In addition, MultiPutRequest was very memory hungry due to the
deeply nested map-of-map-of-map-of-list-of-edits.  This crazy map was
retained during the whole lifetime of the RPC and then fed to the GC.
In the buffering code, edits were added to the map while holding a lock.
Since all ByteMap operations are O(n log n) and there were several of
them involved, the lock was held for long enough to be easily contended
in a write-heavy application.  Just to add even more complexity, the
code was handling durable and non-durable edits separately for no good
reason, which made certain atomic operations on buffered edits more
complicated than is necessary.

The new MultiPutRequest simply aggregates PutRequests together in an
array.  There are far fewer references involved and adding an edit
happens in amortized O(1).  Also we now need to sort the array before
serializing it, and we need an extra O(2N) byte array comparisons, but
this is still far cheaper that the old code which was iterating on the
crazy map.  Plus, overhead at serialization time is a one-time cost
whereas overhead when adding an edit to a MultiPutRequest induces lock
contention, which is way worse as it impedes on scalability.

We now also support single-edit "put" requests.  When we attempt to
send out a single-edit multi-put, it gets automatically transformed
as a single-edit put.

We now also support "unbufferable" edits, which will always be sent
out to HBase directly, regardless of the flush interval.  Previously
this was only possible when an edit was given an explicit lock.  This
is new feature is useful for applications that are mostly throughput
driven but occasionally need low-latency write accesses.

When a MultiPutRequest fails, instead of sending another one we now
split the edits back to single edits and send them individually to
HBase.  Because of HBASE-2898, this is actually more efficient
especially when dealing with a NotServingRegionException, as in this
case it allows us to know precisely which regions are unavailable and
let the edits going to other regions succeed.  It also allows the
edits that aren't affected by the NSRE to proceed immediately.

Incompatible API changes: the following public methods no longer exist:
  HBaseClient.put(PutRequest, boolean)
  HBaseClient.put(List<PutRequest>)
  HBaseClient.put(List<PutRequest>, boolean)
In order to specify whether an edit is to be stored in a durable
fashion, you must now use PutRequest.setDurable(boolean).

Since async HBase is still beta software, I'm taking the liberty to
remove this from API without starting to carry the @Deprecated burden.

Change-Id: I9654e3e39700faaaac19ee9234ffaa60b376bad3