Avoid unnecessary warnings when invalidating the regions cache.
Compare the region names to avoid logging a confusing `Oops' warning
when invalidating a RegionInfo reference that isn't the one we expected
as long as it's for the right region name.

Since the library logs at the INFO level when regions are added to
its local cache, also log cache invalidations are the INFO level to
be consistent.  This can be noisy in large clusters with many regions,
especially when the application starts or when HBase is actively
moving regions around (e.g. due to machine failures).  Maybe we'll
want to tune this down to the DEBUG level (for both messages).

Change-Id: Id56849972740a1b36d26889f9f0f7007a5297c23