Behave better when getting an NSRE during META lookups.
When getting NSRE'd during a META or ROOT access, the code was retrying
properly but it wasn't invalidating the cached entry for META or ROOT.
In some cases, this led to unnecessary failures of RPCs or the code was
not able to trigger a ROOT lookup to trigger a failure of both META and
ROOT.  The code was also not behaving properly when the end-user code
was accessing META or ROOT directly (which doesn't happen in typical
HBase applications).

This changes the GetClosestRowBefore RPC to be retry-able RPC in face
of NSRE failures.  This RPC wasn't going to the normal NSRE handling
code path because it pretended to not be directed to a specific region.

Change tested against a modified RegionServer that throws an NSRE with
a 20% probability (which helped uncover a lot of corner cases).

Change-Id: I67277a75a8b59d8cc650677a8569ad3326e96445