Prevent a deadlock during shutdown.
When shutting down, we properly disconnect all our RegionClients.
After doing so, we check that there are no clients left, and if
we manage to find some, we log an error.  This innocuous logging
statement could lead to a deadlock in rare error cases due to
inconsistent lock ordering.  The required lock ordering is now
explicitly documented and we no longer log the error while
holding the lock, we instead make a copy of the HashMap and then
log the copy.

Change-Id: Ic698dbefb19974f58d6ed534498043e0b83aaf20