Don't flood the logs when -ROOT- is offline.
When -ROOT- was offline, every single RPC that needed to access it was
logging an INFO message.  This problem rarely happened because -ROOT-
is very rarely accessed, but during or after an HBase outage, this
message could be logged for every single RPC.

We now log the message only once, the first time we realize we need to
find where -ROOT- is.

Also, only log a warning when getting disconnected from ZooKeeper while
we're trying to locate the -ROOT- region.  If we're not trying to locate
the -ROOT- region, we don't care about losing our ZooKeeper session.

Change-Id: Iea0b639b91940c4cb4909b8221c91c07f7815079