Schedule periodic flush timers only on demand.
Any RegionClient that was connection to a RegionServer has a timer to
periodically flush buffered edits.  The timer was re-scheduled as long
as the client lived, from the point where it was connected.

This change schedules the timer only when we create a new batch of
buffered edits.

The consequence of this change is that applications that aren't write
heavy, or only have sporadic write periods, or only write to a small
subset of all the RegionServers they use, will have far fewer timers.

Change-Id: Ie793ad6cc79343ec006cccf2e62a07e7d5338407