Fix an NPE caused by a race condition when sending RPCs.
This NPE could occur when a RegionClient was getting disconnected from
its RegionServer while serializing an RPC.  We now double check that
we're still connected before sending the RPC out to the wire.  If we're
not, we let the RPC go through the normal "oops, we're disconnected"
code path.

The NPE could escape out of asynchbase and affect code that uses the
library.

Change-Id: I288a85eedf4ad2fd6922682be67533ede1da52f1