Bug fix: fetch the -ROOT- region from ZK if our client is dead.
The code was sometimes incorrectly using a stale reference for the
client no longer connected to the -ROOT- region.  This would especially
happen when HBase restarts and the client was reading a stale znode in
ZK, depending on the order in which RegionServers are restarted.  When
this bug was triggered, the client was unable to find the new location
of -ROOT- (and thus .META.), and was essentially stuck forever.

Change-Id: I27d2836103a04e4886cd2b92a302c30449f77fbd