Add support for CDH3b3.
This change adds support for CDH3 b3.  In order to enable the new code,
the JVM must be given the following system property in argument:
  -Dorg.hbase.async.cdh3b3

CDH3b3 includes a temporary patch that changes the format of the "hello"
header that clients must send when they connect.  The change is not
backwards compatible and results in clients getting disconnected as
soon as they send the header, because the HBase RPC protocol provides no
mechanism to send an error message back to the client during the initial
"hello" stage.  Thus, older clients will never be able to get past the
initial -ROOT- lookup.

For reference, the patch I'm referring to is:
  74542880d740e9be24b103f1d5f5c6489d01911c
  CLOUDERA-BUILD. HBase running on secure hadoop, temporary patch.

  This is not upstreamed, since it currently is very difficult to do this
  without reflection or a shim layer. This will be upstreamed with the
  larger project of HBase security later this year.

Change-Id: I421e89447e3b55b3000e4ebc486bc90abcbb7076