Bump up Scanner.DEFAULT_MAX_NUM_ROWS from 16 to 128.
16 turned out to be too conservative.  Document the fact that this
setting has a high performance impact on the scanning performance.
A higher default value will provide better performance out of the
box.  OpenTSDB sees a 40-50% speedup with this value.  People with
very large rows will probably be aware of these considerations and
are more likely to adjust the value than those who're using typical
HBase rows, which tend to be fairly limited in size.

Change-Id: I778e28c9211e2c0d462d66f7dd35f0c9ddd5afa6