Make `MultiPutRequest.toString' cheaper and call it less.
When a `MultiPutRequest' contains a lot of edits, it's too expensive to
serialize them all to ASCII for the `toString' representation.  This
adds an upper bound on the amount of data the `toString' representation
will produce.  It also changes `NotServingRegionException', which was a
frequent caller of this method, to lazily generate the message of the
exception, as it is rarely needed anyway.

This should result in fewer CPU cycles wasted and less strain on the
GC when we get an "NSRE storm".  This optimization is warranted by
actual observations that "NSRE storms" would indeed cause significant
CPU consumption and excess GCing due to the problem above.

This closes #3.

Change-Id: I7864c36aa4ef36fdaed51fb0837b0c9e00601453