Add various interfaces to be able to introspect RPCs.
All public RPC types now implement one or more interfaces (such as
`HBaseRpc.HasKey' for RPCs that contain a row key) to provide public
APIs to access the contents of the RPCs.  This is especially important
for unit tests, where client code needs to verify the contents of RPCs
without sending them out over the wire to a server.  This closes #9.

Change-Id: I7b41a6affa12e912e973b533d57912a2bd3df673