Look for -ROOT- again after 1s in case the znode was invalid.
When reading a znode that was corrupt/invalid/empty, the code was simply
giving up.  This is bad because it's not possible to kick off another
lookup because the ZK session will stay up.  Instead keep the session
open and kick off another lookup after a small delay.  Ideally this
should use a ZK watch.

Change-Id: Ibe9246cdac3727bd4c631e9f0e7f4a50e8be4cb0