0.91 compatibility: Change the way the client says hello to the server.
The HBase RPC protocol suffers yet another backward incompatible change
that can't be gracefully detected by clients.  If a 0.90 client connects
to a 0.91 server, unlike with previous versions of HBase where the
client would get unexpectedly disconnected from the server, with 0.91
the connection succeeds but it's in a permanently broken state.  Any RPC
sent to the server will trigger a NPE inside the server.

Add a new JVM flag to enable 0.91 compatibility mode:
  -Dorg.hbase.async.091=true
Once 0.91 becomes the stable version, this flag will be set to true by
default, and clients will have to set it to false in order to revert to
the 0.90 behavior.

Change-Id: Ia4dd23986b2e1a32433915eff5fe0ede3c024dcd