Don't leak version requests.
When sending out the first RPC, we piggy back a getProtocolVersion
request in the first TCP packet.  If we found out that another thread
already sent it, we used to leak the extraneous getProtocolVersion
request that had been created.  So re-order operations to make sure
this doesn't happen.

Bug reported by Jonathan Payne.

Change-Id: Ia464fc02fc8e3457f1231b11d21349048e585b8e