Fix a race condition when sending out the first packet to HBase.
When sending out the first RPC, we need to prepend it with a small
header, which is our way to say "hello" to the RegionServer.  The
code doing this wasn't thread-safe, and so sometimes it was possible
that an RPC would be sent out before the "hello" header.  This would
cause the RegionServer to abruptly disconnect the client.

Change-Id: I49dd1c45fcb8300a2f737ba14662567c8d31d0e2
Signed-off-by: Benoit Sigoure <tsuna@stumbleupon.com>