Fix a race condition that can happen when queuing RPCs.
If a connection to a RegionServer was getting connected/disconnected in
rapid succession, some RPCs could get stuck in a buffer for pending RPCs
and never be sent out.  Calling shutdown() wouldn't flush them either.

The bug could happen with the following sequence of events:

  Thread A                      Thread B
  enter channelConnected()      ender sendRpc()
                                realize that `chan' is null
  set `chan' to non-null
  acquire lock on `this'        block when trying to lock `this'
  swap out pending_rpcs
  release lock on this          acquire lock on `this'
  flush RPCs                    re-create pending_rpcs and add to it

The fix is simply to have Thread B simply re-check whether `chan' is
null after acquiring the lock on `this'.

Bug reported by Jonathan Payne.

Change-Id: I616203b68881efc5d634dfcfeb5589c67ae8a337