Adjust the way we say "hello" to the server to work with 0.92.
Because HBase 0.92 completely broke wire-level compatibility (mostly
because of HBASE-3939), we can no longer start sending any RPC as soon
as we connect.  We must first find out what protocol version the server
is using.

This also fixes an annoyance that could happen when sending rapidly a
lot of version-sensitive RPCs on a newly established connection.  The
previous code could easily reach the limit of number of callbacks on
a Deferred, which was annoying to handle.  The new code doesn't have
this problem because the connection is not deemed usable until we know
what version the server is running.

Change-Id: I45886f0cdf8e29f4a8ce8e99fa0fd71500376128