Overhaul multi-put to support the `multi' RPC.
In HBase 0.92, support for the `multiPut' RPC was removed, as this RPC
method had been deprecated during all of 0.90's lifetime already, and
was superseded with the `multi' method, which also supports batching
`get' and `delete' RPCs.

Refactor all the `multiPut' code to also support the `multi' RPC,
which is even more insanely complicated to serialize due to extra
junk added to what `multiPut' was doing.

Although the `multi' RPC was already supported in 0.90, we don't use
it when talking to 0.90 servers as the compatibility of `multi' between
0.90 and 0.92 was completely broken (unsurprisingly) and supporting both
flavors of `multi' RPCs would make the code uglier than it already is.

Right now `multi' is only used to batch `put' requests.  Another
last step is needed to allow batching `get' and `delete' requests
when talking to 0.92 servers.

Change-Id: I00c0d00b7e90a9b8a1f6987a39592684d981d43d