Fix a race condition while opening new connections.
Previously, as soon as our socket got connected to the RegionServer, we
started using it to send RPCs.  But as of 2a24c7c7, we can no longer do
this as we need to wait until we know what protocol version the server
speaks.  The bug in that change was to allow the RegionClient to start
sending RPCs as soon as it was connected by setting `this.chan' too
early.  Now we set `this.chan' only once we know the server's version,
so that RPCs that are sent in the mean time will get queued to wait.

Change-Id: I65f505c63d21e08d39ab25b556f0a48dbf75047f