Add support for batching deletes.
If we're talking to 0.92, then we can batch deletes (and gets) along
with puts.  So add support for automatically batching `DeleteRequest'
using the `multi' RPC.  Batching of gets might be implemented later,
although it's quite a bit more challenging as HBase's `Get' object
is serialized completely differently, where as `Put' and `Delete' are
very similar (with only a couple of subtle differences, otherwise it
would have been too easy, of course).

Change-Id: I8a5f94f979ba88abecb5b7ed90d6df13512eb85c