Atomic increment coalescing.
Add a new method, `bufferAtomicIncrement', to allow client-side
coalescing of atomic increments.  This closes #20.

In high-throughput user-facing applications that do a lot of counter
increments with a reasonable working set, this can significantly reduce
the number of RPCs and write load on HBase.  For example, at StumbleUpon
we have user-facing application servers that increment various counters
in HBase to keep track of user activity and URL activity, and increment
coalescing is able to reduce 30000 increments/s down to 2000 RPC/s,
a 15x reduction in write load.

Change-Id: I76e427430258d4281c2d4342ff3469d39a8f71a0