Keep track of usage statistics.
Add counters to keep track of various types of activities (RPCs sent,
connections created, etc.) and provide an API to get a snapshot of the
stats.  All counters are using the JSR 166e sharded counter in order
to minimize the overhead of stats collection.

In practice all application servers that use asynchbase need to collect
these stats anyway to expose them to the monitoring system, so we may
as well have the stats built into the client.  At the moment there is
no mechanism to disable stats collection.

Change-Id: I3913b67de69cf32cab01f8aecee392d743f33e62