Make `ensureTableFamilyExists' and `ensureTableExists' cheaper.
Due to a bug in HBase (HBASE-3170), checking whether a table or
table+family exists was actually downloading the first key of the
table (regardless of what the key is).  Instead of doing a `get'
we now only do an `exists' RPC, which still forces the RegionServer
to load the first row in memory, but at least doesn't cause it to
send it over the wire.

Change-Id: Ie26a91e58722b73a6b39804e5d453e5e16b14e77