Allow timestamp=0 on KeyValues.
HBase considers 0 like a legit timestamp, and there is no reason why it
would not.  It was also pointed out that for those who don't want to use
timestamps, setting it to 0 on all their KeyValues helps achieve better
compression.  Some experimental evidence collected on a table with 37B
KeyValues at StumbleUpon shows that this is true indeed, as we've seen
up to 10% better compression in a case where an HBase replication bug
was setting many KeyValues to the same timestamp (see HBASE-6239).

This closes #28.

Change-Id: I9b8d2fd85d4964e0ab7aa28799e8bb406a3ef0e9