Fix serialization bug involving batched puts and deletes.
A `multi' RPC that contains a DeleteRequest for a whole row could lead
to an incorrectly serialized RPC.  This results in the client throwing
a "DeferredGroupException: At least one of the Deferreds failed" with
the first exception being a "RemoteException: IPC server unable to read
call parameters: Error in readFields".  On the RegionServer side, the
log complains about the same "Error in readFields", most likely caused
by a "NegativeArraySizeException: -1" thrown from the `readAttributes'
method of `OperationWithAttributes'.

This bug can only be triggered when writing to HBase 0.92 or above.

This closes #25.

Change-Id: I1bfe41a84abe7a6b4dc45f111e9a504ec8258d1d