Have flush() take into account buffered increments in its return value.
Calling client.flush().join() would correctly flush all buffered atomic
increments, but would return too early, potentially before all the RPCs
doing increments have completed. This was due to the fact that when the
increment buffer gets flushed,  the RPCs sent out weren't tied together
to the Deferred returned.

Getting access to the Deferreds of the atomic increments sent out as the
result of a flush() is hard, because these Deferreds are created from a
callback invoked by Guava upon clearing the increment_buffer.  Instead,
to workaround this problem, call sync() on the RegionClients, a method
introduced to provide a semantic similar to Linux's sync(2) system call.

This closes #38.