Update an atomic increment integration test to be more reliable.
Because these atomic increments can't be buffered, they go out
immediately, and client.flush() may not flush anything (because
the RPCs are have already been sent).  This is working as intended
but it makes the test flaky.  So make sure we really only verify
the final count after the two increments have completed.