Fix the code used to serialized WritableByteArrayComparable.
The code was wrong but worked by chance because the second
code was set to 0 (NOT_ENCODED), and HBase actually uses the
second code to know what type the object really is.

So this is more of a "cosmetic issue on the wire", but better
fix it anyway.

Problem reported by Viral Bajaria.