Allow RPCs to fail-fast.
Add a new pair of public APIs to HBaseRpc to allow specifying
whether an RPC should fail as soon as it encounters a problem.
This is useful for applications with low-latency requirements
where retries are unnecessary anyway.

Once HBase has the ability to track a timeout per RPC this will
be less useful, but in the interim it's a very useful and small
addition.