Refactor how RPC objects know what RPC method they're for.
Introduce a new internal method to HBaseRpc, method(version), which
must return the RPC method name to invoke, given the HBase protocol
version of the server we're connected to.

Previously, all HBaseRpc instances had a `method' byte array in the
object that carried the name of the method to invoke (e.g. "get" or
"put").  Although the names were only created once (statics) this
still added one reference to each and every RPC object.

With HBase 0.95+ compatibility, many RPCs have changed their names
and thus it makes sense to standardize on an internal API that RPCs
must have to declare the name of the RPC method they should invoke
as a function of the HBase protocol version.