Compile Java code generated from protobuf files separately.
There is so much code in there that's a bit slow to compile,
even though javac is far from being a real compiler.  Since
the generated protobuf can be compiled on its own and is less
likely to change (as it would only change as a result of an
update made to one of the .proto files), building it on its
own allows subsequent incremental builds to stay just as fast
as they used to be.