Properly flush queued up RPCs upon connecting.
While we are in the process of connecting, we temporarily queue up RPCs.
Once the connection completes, we flush them.  We used to do so one by
one, without going through the normal code path for sending RPCs.  This
had two problems:
  - We weren't giving the batchable RPCs a chance to get batched.
  - There was a check-then-act sequence that was unsafe and could lead
    to us attempting to write to a closed channel.

This bug would manifest itself as `regression2' failing on a cluster
where the RS serving the test table is different from that serving the
META table (in other words this issue wouldn't be uncovered while using
a non-multi-node HBase cluster).