Fix an edge case in HBase 0.96 one-shot scanners.
This bug was introduced in b0113971: with HBase 0.95+ we have the
ability to start retrieving results in the same RPC as the one opening
the scanner.  However, if that RPC returns nothing because the scanner
didn't match any row in the first region, we invoked the callback with
an empty list of results in argument, which is not supposed to happen.

Factor out some code to properly handle scanners reaching the end of
their region, even if that happens during the first RPC.