Fix accesses to "hbase:meta".
In HBase 0.96, accesses to "hbase:meta" must be treated like accesses to
"-ROOT-" in HBase 0.94 and earlier.  The code responsible for locating
the region of a given table/key pair was missing a special case test for
HBase 0.96.

The code doing .META. lookups was also referring to the old table name,
and while this didn't matter with HBase 0.96 because the name isn't used
in the RPC's serialization, it caused some confusion while debugging.