Add tests to reproduce AlreadyNSREd bug #81
Added two tests to reproduce the bug that will lead to retrial
RPCs when the code goes through KnownToBeNSREd region path in
the sendRpcToRegion in the HBaseClient.  This will turn worse
if the region is NSREd for some good amount time leading to
the failure of probe RPC with TooManyAttempts.

Signed-off-by: Benoit Sigoure <tsunanet@gmail.com>